Align redirect auth drop semantics and split tests
