+++ Docs/reviewer/configuration.md
@@ -65,7 +65,7 @@ This provider uses `POST /v1/chat/completions` and requires:
 Security note: `baseUrl` must use `https://` for non-loopback hosts by default. Plain `http://` is allowed only for loopback
 (for example `http://localhost:11434`). To allow non-loopback `http://`, set `review.openaiCompatible.allowInsecureHttp: true` and `review.openaiCompatible.allowInsecureHttpNonLoopback: true` (dangerous).
 
-Compatibility note: by default, Authorization is kept on same-host redirects. To force-drop it on redirects, set `review.openaiCompatible.dropAuthorizationOnRedirect: true` (or `OPENAI_COMPATIBLE_DROP_AUTHORIZATION_ON_REDIRECT=1`).
+Compatibility note: by default, Authorization is kept on same-host redirects. To force-drop it when a redirect changes scheme/host/port, set `review.openaiCompatible.dropAuthorizationOnRedirect: true` (or `OPENAI_COMPATIBLE_DROP_AUTHORIZATION_ON_REDIRECT=1`).
 
 ```json
 {
